+++ contracts/UtilitiesTests.sol
@@ -17,7 +17,7 @@ contract UtilitiesTests {
     /**
     * @dev The constructor sets the owner
     */
-    constructor ()  public {
+    constructor() public {
         owner = msg.sender;
     }
 
@@ -26,19 +26,19 @@ contract UtilitiesTests {
     *@param _TellorMasterAddress is the Tellor Master address
     */
     function setTellorMaster(address payable _TellorMasterAddress) public {
-        require (msg.sender == owner, "Sender is not owner");
+        require(msg.sender == owner, "Sender is not owner");
         tellorMasterAddress = _TellorMasterAddress;
         tellorMaster = TellorMaster(_TellorMasterAddress);
     }
 
-    function testgetMax() public view returns(uint _max, uint _index) {
+    function testgetMax() public view returns (uint256 _max, uint256 _index) {
         uint256[51] memory requests = tellorMaster.getRequestQ();
-        (_max,_index) = Utilities.getMax(requests);
+        (_max, _index) = Utilities.getMax(requests);
     }
 
-    function testgetMin() public view returns(uint _min, uint _index) {
+    function testgetMin() public view returns (uint256 _min, uint256 _index) {
         uint256[51] memory requests = tellorMaster.getRequestQ();
-        (_min,_index) = Utilities.getMin(requests);
+        (_min, _index) = Utilities.getMin(requests);
     }
 
 }
